Refactoring of Read / Write interfaces and naming conventions Change log: * Added continuous integration   * Formatting, documentation, static analysis   * Compilation and testing with GCC and Clang on Linux with FFTW3 and MPICH. * Refactored trace metadata storage classes. Now called `Trace_metadata`. * Refactored low-level I/O class, now called `Binary_file`, with an MPI implementation called `MPI_Binary_file`. * Refactored file reader and writer interfaces, removing the need for an object layer, making specialization for new seismic file types more obvious. * Renamed namespaces to use all lower-case. * Added a number of static analysis checks, including the cases of variables and functions.

+++ scripts/format.sh
@@ -24,13 +24,22 @@ cat "${file}" \
     | sed 's|//@#pragma omp|#pragma omp|' \
     > "${tmpfile}"
 
+if [ ! -s "${tmpfile}" ]
+then
+    echo "Error! Formatted file is empty!"
+    exit 1
+fi
+
+
 # Swap file for tmpfile, but only if there's a change
 set +o errexit
 if cmp --quiet "${tmpfile}" "${file}"
 then
+    set -o errexit
     echo "Formatting ${file}: No Change..."
     rm "${tmpfile}"
 else
+    set -o errexit
     echo "Formatting ${file}: Updating..."
     mv "${tmpfile}" "${file}"
 fi
